oracle/oracle.go
package oracle

import (
	"errors"
	"strings"

	"github.com/mstgnz/sqlporter"
)

type Oracle struct {
	schema *sqlporter.Schema
}

// NewOracle creates a new Oracle parser instance
func NewOracle() *Oracle {
	return &Oracle{
		schema: &sqlporter.Schema{},
	}
}

// Parse parses Oracle dump content
func (o *Oracle) Parse(content string) (*sqlporter.Schema, error) {
	if content == "" {
		return nil, errors.New("empty content")
	}

	// TODO: Oracle dump parsing operations will be implemented here
	return o.schema, nil
}

// Generate generates Oracle dump from schema
func (o *Oracle) Generate(schema *sqlporter.Schema) (string, error) {
	if schema == nil {
		return "", errors.New("empty schema")
	}

	var result strings.Builder

	// TODO: Oracle dump generation operations will be implemented here

	return result.String(), nil
}
